refactor(Media): ♻️ Simplified figure component & enhanced image handling.     -	Removed redundant `isFigure` checks in `Media/Figure.vue` to     simplify component logic.
    -	Implemented a fallback mechanism that uses the `alt` text for
    images when a caption is not provided.

    -	Renamed the `image` property to `media` in `Media/Figure.vue` to
    maintain consistency across the codebase.

    -	Introduced conditional checks for `isImageModern` and
    `isImageStandard` within the `picture` element to support various
    format strategies.

    -	Updated all type references in both `Media/Figure.vue` and
    `Media/Icon.vue` to align with the latest interface definitions.
